requirements.txt
Flask-Bootstarp4
numpy
pandas
psycopg2
python-dotenv